Refatroando o sistema
