+++ docker-compose.yml
@@ -1,0 +1,10 @@
+version: '3.8'
+services:
+  sql-server:
+    build: .
+    container_name: sql-server-container
+    environment:
+      SA_PASSWORD: rael2438$
+      ACCEPT_EULA: Y
+    ports:
+      - "1533:1433"
