+++ dockerfile
@@ -1,0 +1,12 @@
+FROM mcr.microsoft.com/mssql/server:2019-latest
+
+ENV ACCEPT_EULA=Y
+ENV SA_PASSWORD=rael2438$
+ENV MSSQL_PID=Developer
+ENV MSSQL_TCP_PORT=1433
+
+WORKDIR /src
+
+COPY scripts.sql ./scripts.sql
+
+RUN (/opt/mssql/bin/sqlservr --accept-eula & ) | grep -q "Service Broker manager has started" &&  /opt/mssql-tools/bin/sqlcmd -S127.0.0.1 -Usa -Prael2438$ -i scripts.sql
