+++ src/CustomerManager.Web/CustomerManager.Web.csproj
@@ -11,4 +11,9 @@
     <PackageReference Include="Microsoft.Data.SqlClient" Version="5.2.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CustomerManager.Core\CustomerManager.Core.csproj" />
+    <ProjectReference Include="..\CustomerManager.Infrastructure\CustomerManager.Infrastructure.csproj" />
+  </ItemGroup>
+
 </Project>
