+++ src/CustomerManager.Web/Models/AddressListView.cs
@@ -1,0 +1,10 @@
+﻿using CustomerManager.Core.Models;
+
+namespace CustomerManager.Web.Models
+{
+    public class AddressListView
+    {
+        public int CustomerId { get; set; }
+        public IList<Address> Addresses { get; set; } = null!;
+    }
+}
