--- src/CustomerManager.Web/Models/Customer.cs
@@ -1,29 +1,0 @@
-﻿using CustomerManager.Web.Commons;
-using CustomerManager.Web.Commons.Utils;
-using System.ComponentModel.DataAnnotations;
-
-namespace CustomerManager.Web.Models
-{
-	public class Customer
-	{
-		[Key]
-		public int Id { get; set; }
-
-		[Required(ErrorMessage = "Campo 'Nome Completo' é obrigatório")]
-		public string Name { get; set; } = string.Empty;
-
-		[Required(ErrorMessage = "Campo 'Telefone' é obrigatório")]
-		public string Phone { get; set; } = string.Empty;
-
-        [CpfValidate]
-        public string CPF { get; set; } = string.Empty;
-
-		[Required]
-		public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-
-		[Required]
-		public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
-
-		public DateTime? DeletedAt { get; set; } = null;
-	}
-}
