+++ src/CustomerManager.Web/Models/CustomerEditFormViewModel.cs
@@ -1,0 +1,10 @@
+﻿using CustomerManager.Core.Models;
+
+namespace CustomerManager.Web.Models
+{
+    public class CustomerEditFormViewModel
+    {
+        public Customer? Customer { get; set; } = null!;
+        public IEnumerable<Address> Addresses { get; set; } = null!;
+    }
+}
