--- src/CustomerManager.Web/Models/ViewModels/AddressListView.cs
@@ -1,8 +1,0 @@
-﻿namespace CustomerManager.Web.Models.ViewModels
-{
-    public class AddressListView
-    {
-        public int CustomerId { get; set; }
-        public IList<Address> Addresses { get; set; } = null!;
-    }
-}
