--- src/CustomerManager.Web/Models/ViewModels/CustomerEditFormViewModel.cs
@@ -1,8 +1,0 @@
-﻿namespace CustomerManager.Web.Models.ViewModels
-{
-    public class CustomerEditFormViewModel
-    {
-        public Customer? Customer { get; set; } = null!;
-        public IEnumerable<Address> Addresses { get; set; } = null!;
-    }
-}
