+++ src/CustomerManager.Web/Views/Customer/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@using CustomerManager.Web.Models.ViewModels
+﻿@using CustomerManager.Web.Models
 @model CustomerEditFormViewModel
 @{
     ViewData["Title"] = "Editar Cliente";
