+++ src/CustomerManager.Web/Views/_ViewImports.cshtml
@@ -1,4 +1,5 @@
 ﻿@using CustomerManager.Web
 @using CustomerManager.Web.Models
-@using CustomerManager.Web.Commons.Helpers;
+@using CustomerManager.Web.Helpers;
+@using CustomerManager.Core.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
