+++ src/CustomerManager.Web/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "AppDataConnectionStr": "Data Source=(LocalDb)\\MSSQLLocalDB;Integrated Security=SSPI;AttachDBFilename=%CONTENT_PATH%\\Data\\customer_db.mdf"
+    "AppDataConnectionStr": "Server=127.0.0.1,1533;Initial Catalog=customer_db;User Id=sa;Password=rael2438$;"
   }
 }
